Move time_delta function from process.py to utilities.py
importmail.py was trying to import process.py for the time_delta function, but failed because process.py requires TK (see #450).
